requirements.txt
asgiref==3.9.1
Django==5.2.5
django-filter==25.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
promise==2.3
pycodestyle==2.14.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.1
